refactor css files and respective jsx code
