+++ src/App.js
@@ -1,17 +1,19 @@
-import React, { Component } from 'react';
+import React, {Component} from 'react';
 
 import Home from './scenes/Home'
 import Footer from './components/Footer';
+import Header from "./components/Header";
 
 class App extends Component {
-  render() {
-    return (
-      <div className="App">
-        <Home />
-        <Footer />
-      </div>
-    );
-  }
+    render() {
+        return (
+            <div className="App">
+                <Header/>
+                <Home/>
+                <Footer/>
+            </div>
+        );
+    }
 }
 
 export default App;
