+++ src/assets/logo_secondary.png
