--- src/components/Footer/styles.scss
@@ -1,33 +1,0 @@
-.footer {
-    background-color: #1c0f09;
-  
-    .links {
-      align-items: center;
-      display: flex;
-      flex-direction: row;
-      height: 96px;
-      justify-content: center;
-  
-      a {
-        border: 2px solid white;
-        border-radius: 100px;
-        height: 60px;
-        margin: 0 15px;
-        width: 60px;
-  
-        img {
-          display: block;
-          height: 26px;
-          margin: auto;
-          padding-top: 16px;
-          position: relative;
-        }
-      }
-    }
-    .lower-links {
-      text-align: center;
-      padding-bottom: 20px;
-      
-      a, p { display: inline; color: white; font-size: 18px; }
-    }
-  }
