+++ src/components/Header/index.jsx
@@ -1,0 +1,8 @@
+import React from 'react';
+
+const Header = () => (
+    <div className="header">
+    </div>
+);
+
+export default Header;
