--- src/index.scss
@@ -1,16 +1,0 @@
-.App {
-  height: 100vh;
-}
-
-* {
-  font-family: 'Montserrat', sans-serif;
-  margin: 0;
-  padding: 0;
-}
-
-@font-face {
-  font-family: 'Dust West';
-  font-style: normal;
-  font-weight: normal;
-  src: local('MyFont'), url('./assets/fonts/Dust West.woff') format('woff');
-}
