+++ src/scenes/Home/index.jsx
@@ -1,17 +1,16 @@
-import React, { Component } from 'react';
+import React, {Component} from 'react';
 
-import MainView from './scenes/MainView.jsx';
-import About from './scenes/About.jsx';
-import Sponsors from './scenes/Sponsors.jsx';
-import './styles.scss';
+import About from './components/About.jsx';
+import Sponsors from './components/Sponsors.jsx';
+import Logo from '../../assets/logo_main.png';
 
 class Home extends Component {
     render() {
         return (
             <div className="home">
-                <MainView />
-                <About />
-                <Sponsors />
+                <img className="logo" src={Logo} alt="logo"/>
+                <About/>
+                <Sponsors/>
             </div>
         );
     }
