--- src/scenes/Home/styles.scss
@@ -1,160 +1,0 @@
-.splash, .contact-wrapper {
-    background-image: url('../../assets/backdrop.jpg');
-    background-size: cover;
-    background-position: center center;
-}
-
-h1 {
-    font-family: 'Dust West';
-    font-size: 70px;
-    color: #e3af40;
-    margin-bottom: 30px;
-}
-
-.splash {
-    width: 100%;
-    height: 100vh;
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    justify-content: center;
-
-    #main_logo {
-        width: 60%;
-    }
-}
-
-.about-wrapper {
-    background-color: #1c0f09;
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    justify-content: center;
-
-    padding: 90px 0 80px 0;
-
-    #slogan {
-        width: 40%;
-        margin-bottom: 60px;
-    }
-
-    .about-text {
-        color: #ebca97;
-        padding: 0 170px 0 170px;
-
-        p {
-            font-size: 22px;
-            margin-bottom: 20px;
-        }
-    }
-}
-
-.contact-wrapper {
-    height: 60vh;
-    text-align: center;
-    padding-top: 80px;
-    
-}
-
-.sponsors-wrapper {
-    background: #756249;  /* fallback for old browsers */
-    background: -webkit-linear-gradient(to right, #756249, #61523d, #403628);
-    background: linear-gradient(to right, #756249, #61523d, #403628);
-    
-    text-align: center;
-    padding: 80px 0 60px 0;
-
-    p {
-        text-align: left;
-        margin: 0px 140px 30px 140px;
-        font-size: 20px;
-        color: #ebca97;
-    }
-}
-
-.sponsor-logos {
-    background-color: #c9b391;
-    box-sizing: border-box;
-    box-shadow: -2px 2px 15px #1c0f09;
-    padding: 70px 80px 30px 80px;
-
-    .logo {
-        margin-bottom: 60px;
-    }
-
-    .logos1, .logos2, .logos3 {
-        display: flex;
-        flex-direction: row;
-        flex-wrap: wrap;
-        justify-content: center;
-        align-items: center;
-    }
-    .logos1 {
-        img {
-            width: 100%;
-            max-width: 400px;
-            margin: 0 25px;
-        }
-    }
-    .logos2 {
-        img {
-            width: 100%;
-            max-width: 340px;
-            margin: 0 55px;
-        }
-    }
-    .logos3 {
-        img {
-            width: 100%;
-            max-width: 500px;
-        }
-    }
-    
-}
-
-// Container styling
-$container-widths: (
-  sm: 540px,
-  md: 720px,
-  lg: 960px,
-  xl: 1140px
-);
-
-$grid-breakpoints: (
-  xs: 0,
-  sm: 576px,
-  md: 768px,
-  lg: 992px,
-  xl: 1200px
-);
-
-$narrow-container-widths: (
-  sm: 520px,
-  md: 700px,
-  lg: 850px,
-  xl: 990px 
-);
-
-.container {
-  box-sizing: border-box;
-  margin: auto;
-  padding-left: 15px;
-  padding-right: 15px;
-  width: 100%;
-
-  @each $size, $width in $container-widths {
-    $break: map-get($map: $grid-breakpoints, $key: $size);
-    @media (min-width: $break) {
-      max-width: $width;
-    }
-  }
-}
-
-.narrow-container {
-  @each $size, $width in $narrow-container-widths {
-    $break: map-get($map: $grid-breakpoints, $key: $size);
-    @media (min-width: $break) {
-      max-width: $width;
-    }
-  }
-}
