+++ src/styles/colors.scss
@@ -1,0 +1,9 @@
+$eweek-black: #1c0f09;
+$eweek-gray: #524534;
+$eweek-brown: #3f1f14;
+$eweek-tan: #743f1d;
+$eweek-yellow: #e3af40;
+$eweek-beige: #ebca97;
+$eweek-pink: #e7b092;
+$eweek-red: #782a1d;
+$eweek-blue: #068fdf;
