+++ src/styles/main.scss
@@ -1,0 +1,4 @@
+@import "colors";
+@import "app";
+@import "home";
+@import "footer";
